Make test more deterministic.
